Refactored typescript template into a "wrapper" template
