+++ packages/realm/package.json
@@ -89,6 +89,7 @@
     "build:ts": "wireit",
     "build:node": "wireit",
     "bindgen:jsi": "wireit",
+    "bindgen:wrapper": "wireit",
     "check-types": "wireit",
     "check-circular-imports": "wireit",
     "install": "prebuild-install --runtime napi || echo 'Failed to download prebuild for Realm'",
@@ -201,6 +202,21 @@
       ],
       "output": [
         "binding/jsi/jsi_init.cpp"
+      ]
+    },
+    "bindgen:wrapper": {
+      "command": "realm-bindgen --template bindgen/src/templates/wrapper.ts --spec bindgen/vendor/realm-core/bindgen/spec.yml --spec bindgen/js_spec.yml --opt-in bindgen/js_opt_in_spec.yml --output ./src/binding || true",
+      "dependencies": [
+        "bindgen:generate:spec-schema"
+      ],
+      "files": [
+        "bindgen/vendor/realm-core/bindgen/spec.yml",
+        "bindgen/vendor/realm-core/bindgen/src",
+        "bindgen/js_spec.yml",
+        "bindgen/js_opt_in_spec.yml",
+        "bindgen/src",
+        "!bindgen/src/templates",
+        "bindgen/src/templates/wrapper.ts"
       ]
     },
     "bindgen:generate:typescript": {
